Integrate Input and Button components into page file
Integrates `Button` and `Input` components directly into `src/app/page.tsx` and removes their separate files.

- Moves the `Button` and `Input` component definitions from their individual files into `src/app/page.tsx`.
- Deletes the `src/app/component/button.tsx` and `src/app/component/input.tsx` files as they are no longer needed.
- Ensures that the integrated components are correctly functioning within the page context by maintaining their original functionality and properties.


---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/geo-yuheng/Website/pull/1?shareId=00e88472-6229-43f4-b2a2-b808979d7b76).
